profile-service/pyproject.toml
[project]
name = "profile-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.31.0",
    "fastapi>=0.133.1",
    "psycopg2-binary>=2.9.11",
    "psycopg[binary]>=3.3.3",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.13.1",
    "pyjwt>=2.11.0",
    "python-dotenv>=1.2.1",
    "python-jose>=3.5.0",
    "sqlalchemy>=2.0.47",
    "uvicorn>=0.41.0",
]
